Add HIGHLIGHT_SELECTED, HIGHLIGHT_IMPACTED and HIGHLIGHT_BACKGROUND in Renderer and CSGNode
To increase the visual feedback when an object is selected we add more states.
HIGHLIGHT represents the object selected in source code witht the # (dash) operator
HIGHLIGHT_SELECTED represents the object selected by mouse selection in the 3D view.
HIGHTLIGHT_IMPACTED represents the objects that will be impacted if the one selected is changed.
HIGHTLIGHT_BACKGROUND is there for moving the background management as a subcase of HIGHLIGHT.